Change ContractMethodCall message payload type from Any to bytes
The contract request dispatcher can determine the correct payload type
from the request content type and the argument type obtained via Go's
reflection, so we don't have to rely on the protobuf lib to figure it
out.

The current C# implementation of the protobuf Any type relies on
reflection and is thus incompatible with Unity AOT builds for WebGL and
iOS... so eliminating the Any type makes it slightly easier to support
those gimped platforms.