Refactor the CLI example
Eventually it'd be nice if parameters to smart contract methods could be
passed in as a JSON object, the problem is doing that via command line
is a pain due to all the double-quotes that need to be escaped. Need to
find a way to pass in plain JS objects instead.