Fix wrong caller address in FakeContext Call/StaticCall
The contract context created by the PluginVM sets the caller address
to the contract address stored in the context before calling the
contract. So when contract A calls contract B the latter would see the
caller is contract A.

FakeContext on the other hand left the caller address unchanged before
calling a contract, so when contract A called contract B the latter
would see the caller is whoever called contract A, rather than
contract A itself.